requirements.txt
flask
openai
gunicorn
Flask-SQLAlchemy
Flask-WTF
python-docx